+++ client/shared/src/main/scala/io/quckoo/client/internal/Driver.scala
@@ -1,0 +1,8 @@
+package io.quckoo.client.internal
+
+/**
+  * Created by alonsodomin on 06/09/2016.
+  */
+trait Driver[P <: Protocol] {
+
+}
